+++ app/src/main/java/com/elegion/tracktor/utils/DistanceConverter.java
@@ -72,7 +72,7 @@ public class DistanceConverter implements IDistanceConverter {
     }
 
     public interface ICurrentPreferences {
-        double getUnit();
+        int getUnit();
 
         List<String> getDistanceUnitSymbol();
